Squashed commit of the following:

commit bac93e1a4ea412ff4c748b815e28260fcaf9d72e
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 14:18:20 2020 +0100

    renamed some variables and removed dead code

commit 3ea24d6bb2856a6b05c76f4e0937bb5d9e7ffdd1
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 14:07:03 2020 +0100

    Removed obsolete function altogether

commit aa41c3fcc990839990dc842d4103fff5162fd9ec
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 11:44:03 2020 +0100

    removed redundant comparison with true

commit c62f91e46eb5f5ae5f61a4bacd17273359e5bc1d
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 11:43:44 2020 +0100

    removed always true false,null

commit fe16520daf5e0c1ed91226007d2db5efcb8f6343
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 11:43:05 2020 +0100

    removed 2 unused variables

commit 9d340e429955537c88a57deaffbead98fb77bbc1
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 11:41:57 2020 +0100

    removed hidden default value

commit 35689e6f61bda67a781fad5f8840274a93c799f6
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 11:41:26 2020 +0100

    removed unused function

commit 8c91732e925d93b5075db25e28b0861cc9949edd
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 10:50:53 2020 +0100

    Use ternary to shorten code

commit 6ca1b2088cfa9320ff460e3800fb315cb4ce94d7
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 10:44:27 2020 +0100

    Remove everything from function, except close and rename to 'TryClose(ByVal brokenStream As System.IO.StreamWriter)'

commit e6868f04fbad2888c0a26cc9f633d87301ee41d6
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 10:22:43 2020 +0100

    Replace with getawaiter.getresult -> so exception does not get wrapped.

commit 955535f4d011f4f65da426a675c827b52c5fc93c
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Thu Jan 2 10:18:36 2020 +0100

    Removed weird using

commit 4c5d6cb1a4006c2a8d4657e66355ab8f48adb132
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Mon Dec 23 14:26:33 2019 +0100

    some additional warnings removed

commit 5a17a4a691e76f55e91e8a18988ce936b82c5033
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Mon Dec 23 14:14:14 2019 +0100

    removed unneeded variable

commit 99508040eb17dea6fb79f1cba30ee81ee5c81ddb
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Mon Dec 23 14:03:47 2019 +0100

    removed unused using

commit 9737a219bda08e71c247b0cfa54395c0a15237de
Author: Roland ten Have <Roland.tenHave@relativity.com>
Date:   Mon Dec 23 13:59:29 2019 +0100

    fixed bunch of warnings in code, and removed some obsolete things

+++ Source/Relativity.DataExchange.Core/Data/DelimitedFileImporter2.cs
@@ -777,7 +777,7 @@ namespace Relativity.DataExchange.Data
 		/// </returns>
 		private bool ValidateAppend(long startPosition, int maxCharacterLength, ref bool hasAlertedError)
 		{
-			if (checked(this.CharacterPosition - startPosition) > (long)maxCharacterLength)
+			if (checked(this.CharacterPosition - startPosition) > maxCharacterLength)
 			{
 				if (!hasAlertedError)
 				{
@@ -870,11 +870,11 @@ namespace Relativity.DataExchange.Data
 			StringBuilder stringBuilder = new StringBuilder();
 			long initialCharacterPosition = this.CharacterPosition;
 			bool hasAlertedError = false;
-			if (this.Peek() == (int)this.Bound)
+			if (this.Peek() == this.Bound)
 			{
 				// The Read character is NOT used.
 				Microsoft.VisualBasic.Strings.ChrW(this.Read());
-				if ((int)Microsoft.VisualBasic.Strings.ChrW(this.Peek()) == (int)this.Delimiter ||
+				if (Microsoft.VisualBasic.Strings.ChrW(this.Peek()) == this.Delimiter ||
 				    this.Peek() == EofChar ||
 				    this.Peek() == 13)
 				{
@@ -921,7 +921,7 @@ namespace Relativity.DataExchange.Data
 			while (this.Peek() != EofChar)
 			{
 				char ch1 = Microsoft.VisualBasic.Strings.ChrW(this.Read());
-				if ((int)ch1 == (int)this.Bound)
+				if (ch1 == this.Bound)
 				{
 					if (this.TrimOption == TrimOption.Both ||
 					    this.TrimOption == TrimOption.Trailing)
@@ -934,7 +934,7 @@ namespace Relativity.DataExchange.Data
 					}
 
 					int num = this.Peek();
-					if (num == (int)this.Delimiter)
+					if (num == this.Delimiter)
 					{
 						this.Read();
 						return this.ConvertNewLine(stringBuilder);
@@ -963,7 +963,7 @@ namespace Relativity.DataExchange.Data
 							break;
 
 						default:
-							if (num == (int)this.Bound)
+							if (num == this.Bound)
 							{
 								this.Append(
 									stringBuilder,
@@ -1026,7 +1026,6 @@ namespace Relativity.DataExchange.Data
 					currentArrayList = new ConditionalArrayList(false);
 				}
 
-				// TODO: Verify the usage of "Or" vs "OrAlso"
 				while ((Microsoft.VisualBasic.Strings.ChrW(charCode) == ' '
 				        || Microsoft.VisualBasic.Strings.ChrW(charCode) == '\t')
 				       && (this.TrimOption == TrimOption.Both
@@ -1036,12 +1035,12 @@ namespace Relativity.DataExchange.Data
 				}
 
 				char ch = Microsoft.VisualBasic.Strings.ChrW(charCode);
-				if ((int)ch == (int)this.Bound)
+				if (ch == this.Bound)
 				{
 					currentArrayList.Add(
 						this.GetBoundedFieldValue(ref hasHitEndOfLine, currentArrayList.Count, maximumFieldLength));
 				}
-				else if ((int)ch == (int)this.Delimiter)
+				else if (ch == this.Delimiter)
 				{
 					// Add an empty value
 					currentArrayList.Add(string.Empty);
